feat: merge-permissions-to-org
