+++ api/org-api/src/index.ts
@@ -1,8 +1,10 @@
-import { BASE_PATH } from "@/lib/constants"
+import { BASE_PATH, PERMISSIONS_BASE_PATH } from "@/lib/constants"
 import { middlewares } from "@/middleware"
 import { routes } from "@/routes"
+import permissionsReferenceRoutes from "@/routes/permissions-reference"
 import type { HonoApp } from "@/types"
 import { createService } from "@incmix-api/utils"
+import { PermissionService } from "@incmix-api/utils/authorization"
 
 import { envVars } from "./env-vars"
 
@@ -12,9 +14,27 @@ const service = createService<HonoApp["Bindings"], HonoApp["Variables"]>({
   basePath: BASE_PATH,
   setupMiddleware: (app) => {
     middlewares(app)
+
+    // Custom RBAC setup that excludes reference endpoints
+    app.use(`${BASE_PATH}/*`, (c, next) => {
+      const path = c.req.path
+      // Skip RBAC for reference/documentation endpoints
+      if (path.includes("/reference") || path.includes("/openapi.json")) {
+        return next()
+      }
+      // Apply RBAC to all other routes
+      c.set("rbac", new PermissionService(c))
+      return next()
+    })
   },
-  needRBAC: true,
-  setupRoutes: (app) => routes(app),
+  needRBAC: false, // Disable default RBAC since we're doing it manually
+  setupRoutes: (app) => {
+    routes(app)
+    
+    // Mount permissions reference routes AFTER main routes and OpenAPI setup
+    // Use mount() instead of route() to truly isolate from main OpenAPI spec
+    app.mount(`${BASE_PATH}/permissions/reference`, permissionsReferenceRoutes.fetch)
+  },
 })
 
 const { app, startServer } = service
