+++ api/org-api/src/lib/constants.ts
@@ -1,6 +1,7 @@
 import { API } from "@incmix/utils/env"
 
 export const BASE_PATH = API.ORG
+export const PERMISSIONS_BASE_PATH = API.PERMISSIONS
 
 export const ERROR_ORG_EXIST = { namespace: "errors", key: "org_exist" }
 export const ERROR_ORG_NOT_FOUND = {
@@ -44,4 +45,28 @@ export const ERROR_ROLE_ALREADY_EXISTS = {
 export const ERROR_ROLE_NOT_FOUND = {
   namespace: "errors",
   key: "role_not_found",
+}
+export const ERROR_PERMISSION_EXIST = {
+  namespace: "errors",
+  key: "permission_exist",
+}
+export const ERROR_PERMISSION_NOT_FOUND = {
+  namespace: "errors",
+  key: "permission_not_found",
+}
+export const ERROR_PERMISSION_CREATE_FAIL = {
+  namespace: "errors",
+  key: "permission_create_fail",
+}
+export const ERROR_PERMISSION_UPDATE_FAIL = {
+  namespace: "errors",
+  key: "permission_update_fail",
+}
+export const ERROR_PERMISSION_DELETE_FAIL = {
+  namespace: "errors",
+  key: "permission_delete_fail",
+}
+export const PERMISSION_DELETE_SUCCESS = {
+  namespace: "auth",
+  key: "permission_delete_success",
 }
