+++ api/org-api/src/routes/index.ts
@@ -7,4 +7,6 @@ import healthcheckRoutes from "./health-check"
 export const routes = (app: OpenAPIHono<HonoApp>) => {
   app.route(`${BASE_PATH}/healthcheck`, healthcheckRoutes)
   app.route(BASE_PATH, orgRoutes)
+  // Permissions and roles routes are removed from main org API
+  // They have their own dedicated OpenAPI documentation at /api/org/permissions/reference
 }
