--- api/permissions-api/README.md
@@ -1,10 +1,0 @@
-# Todo API
-
-```bash
-pnpm install
-pnpm run dev
-```
-
-```bash
-pnpm run deploy
-```
