--- api/permissions-api/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "@incmix-api/permissions-api",
-  "type": "module",
-  "scripts": {
-    "dev": "tsx watch src/index.ts",
-    "build": "tsc && tsc-alias",
-    "start": "node ./dist/index.js",
-    "test": "vitest --run",
-    "clean": "rm -rf dist"
-  },
-  "dependencies": {
-    "@casl/ability": "catalog:",
-
-    "@hono/zod-openapi": "catalog:",
-    "@hono/zod-validator": "catalog:",
-    "@incmix-api/utils": "workspace:^",
-    "@incmix/utils": "catalog:",
-    "hono": "catalog:",
-    "kysely": "catalog:",
-    "kysely-ctl": "catalog:",
-    "lodash": "catalog:",
-    "nanoid": "catalog:",
-    "pg": "catalog:",
-    "zod": "catalog:"
-  },
-  "devDependencies": {
-    "@incmix-api/config": "workspace:^",
-    "@types/lodash": "catalog:",
-    "@types/node": "catalog:",
-    "@types/pg": "catalog:",
-    "kysely-ctl": "catalog:",
-    "tsc-alias": "catalog:",
-    "tsx": "catalog:",
-    "typescript": "catalog:",
-    "vitest": "catalog:"
-  }
-}
