--- api/permissions-api/src/middleware/index.ts
@@ -1,14 +1,0 @@
-import { BASE_PATH } from "@/lib/constants"
-import type { HonoApp } from "@/types"
-import type { OpenAPIHono } from "@hono/zod-openapi"
-import { setupApiMiddleware } from "@incmix-api/utils/middleware"
-import { envVars } from "../env-vars"
-
-export const middlewares = (app: OpenAPIHono<HonoApp>) => {
-  setupApiMiddleware(app, {
-    basePath: BASE_PATH,
-    serviceName: "permissions-api",
-    databaseUrl: envVars.DATABASE_URL,
-    corsFirst: true,
-  })
-}
