--- api/permissions-api/src/routes/health-check/types.ts
@@ -1,8 +1,0 @@
-import { z } from "@hono/zod-openapi"
-
-export const HealthCheckSchema = z
-  .object({
-    status: z.string().openapi({ example: "UP" }),
-    reason: z.string().optional().openapi({ example: "database down" }),
-  })
-  .openapi("Healthcheck")
