--- api/permissions-api/test/basic.test.ts
@@ -1,7 +1,0 @@
-import { describe, expect, it } from "vitest"
-
-describe("Permissions API", () => {
-  it("should have basic test setup", () => {
-    expect(true).toBe(true)
-  })
-})
