--- api/permissions-api/vitest.config.mts
@@ -1,11 +1,0 @@
-import tsconfigPaths from "vite-tsconfig-paths"
-import { defineConfig } from "vitest/config"
-
-export default defineConfig({
-  plugins: [tsconfigPaths()],
-  test: {
-    environment: "node",
-    globals: true,
-    dir: "./test",
-  },
-})
