+++ api/projects-api/src/routes/tasks/index.ts
@@ -58,7 +58,8 @@ tasksRoutes.doc("/openapi.json", {
   info: {
     version: "1.0.0",
     title: "Tasks API",
-    description: "Endpoints for task management. Auth via cookieAuth (session).",
+    description:
+      "Endpoints for task management. Auth via cookieAuth (session).",
   },
   tags: [
     {
